Implement global shortcut to center active window
Using Windows "key hooks" we bind to all keyboard events. This allows
listening for specific keyboard combinations and performing certain
actions when they occur. In this case we've bound CTRL+WIN+G to a
command that finds the current active window and centers it in the
screen. This is all done via Win32 API calls using PInvoke.