Add menu action to generate a new icon
The application icon in the system tray can now be regenerated on the
fly via the context menu.

AppIcon no longer exposes an Icon instance. The only way to get the
icon is to register a listener. This will force all consumers of
AppIcon to treat the application icon as a value that can change over
time, and respond accordingly when it does change.